+++ doc/changelog.rst
@@ -11,6 +11,7 @@ Added
 - :class:`~scim2_models.ServiceProviderConfig`, :class:`~scim2_models.ResourceType`
   and :class:`~scim2_models.Schema` are added to the default resource types list.
 - Any custom URL can be used with all the :class:`~scim2_client.SCIMClient` methods.
+- :class:`~scim2_client.ResponsePayloadValidationError` implementation.
 
 Fixed
 ^^^^^